+++ examples/animation/src/main.rs
@@ -15,11 +15,7 @@ use widgets::{ease_btn, ruler};
 
 fn main() {
     zng::env::init!();
-    zng::app::crash_handler::init_debug();
-    app_main();
-}
 
-fn app_main() {
     APP.defaults().run_window(async {
         Window! {
             title = "Animation Example";