+++ examples/layer/src/main.rs
@@ -12,11 +12,7 @@ use zng::{
 
 fn main() {
     zng::env::init!();
-    zng::app::crash_handler::init_debug();
-    app_main();
-}
 
-fn app_main() {
     APP.defaults().run_window(async {
         // layer will init with the window, when it opens.
         LAYERS.insert(