+++ examples/text/src/main.rs
@@ -17,11 +17,7 @@ mod form;
 
 fn main() {
     zng::env::init!();
-    zng::app::crash_handler::init_debug();
-    app_main();
-}
 
-fn app_main() {
     APP.defaults().run_window(async {
         let fs = var(Length::Pt(11.0));