+++ CHANGELOG.md
@@ -7,15 +7,19 @@
 * Implement `From<Txt>` for `std::ffi::OsString`.
 * **Breaking** Remove `zng_view::init`, `zng_view_prebuilt::init`, `zng::view_process::default::init`, `zng::view_process::prebuilt::init`.
     - Only `zng::env::init!()` needs to be called to setup the view-process.
+* **Breaking** Remove `zng_view::extensions::ViewExtensions::new`.
+    - Now use `zng_view::view_process_extension!` to declare.
 * **Breaking** `zng_view_api::Controller::start` now requires the exe path and supports optional env variables to set.
     - This only affects custom view-process implementers.
 * **Breaking** New default `zng::env::res`, was `./assets` now is `./res`.
 * **Breaking** Renamed `zng_view_prebuilt::ViewLib::init` to `view_process_main`.
 * **Breaking** Remove `is_single_instance`, `single_instance`, `single_instance_named`.
     - Single instance now enabled automatically just by setting `feature="single_instance"`.
-* **Braking** Add name requirement for `zng::task::ipc` workers.
+* **Breaking** Add name requirement for `zng::task::ipc` workers.
     - Workers are no longer limited to a single entry with an enum switch.
     - Use `zng::env::on_process_start!` to declare the worker entry anywhere.
+* **Breaking** Remove `zng::app::crash_handler::init` and `CrashConfig::new`.
+    - Use `zng::app::crash_handler::crash_handler_config` to config.
 
 # 0.6.2